Added specs, module and controllers for Places
I've implemented specs, modules and controllers for Places. I've also
modificated the routes files to hold places resources.

It seems to be a problem with the specs for places since it uses two
fields as floats. I've posted a issue for the people at rspec-rails at
https://github.com/rspec/rspec-rails/issues/1798.

Once I got answer about this issue I'll open a new card to work on its
fix.

https://trello.com/c/43GlDS4i/