Fixes on specs for places and producers
The problem was that I was testing over a method rather than the
response itself. It's quite different to test a `response` than a
`json`. Both have differences. If it's required to use `response`, then
the call must be `response.body` or `json`, for example
`response.body.size` is equal to `json.size`.

https://trello.com/c/sb19gp7b/