Creating RouteLog model, adding validations specs.
I created the model of RouteLog applaing a initial change where
location was define as location type, but is not posible to
implement this gem for ruby 5.0, reason why @amrendonsa an I
decided to replaced for lat:float and lon:float.

I created the factory for RouteLog at ./spec/factories/route_logs.rb

There still presenting errors of validation, currently are 8 errors.

https://trello.com/c/jDiDQplP/33-route-log