Creating model,spec,relations,controller,factory for Package
Ensure the relation between Package with Crop, Route and Itself.
Define the specs and created the migration for column show and
set it as true for default.

The route is define for :3000 /v1/packages but because Package
has a self join the test was trouble with the parent_id.

Currently 14 Fails on test and its not posible to add a package with
http POST :3000/v1/packages crop_id=1 route_id=1 quantity=54635.55

https://trello.com/c/qrdAuMoW/34-package