Fixing definitions of queries for `Producer`
I've pair programmed with https://github.com/annipi

We checked what are the real meaning of the `Producer` queries. Instead
of writing useless queries we refactored the `Producer` controller
methods to left them as dummy as possible delegating all the business
logic to the model.

Now the model deals with all the logic behind the Producer including the
way it abstracts the elimination of a single `Producer`.

https://trello.com/c/k2hYxDre/