Fixes on models to have tests passing
We are dealing with a strange behavior on rails for polymorphic
relationships and STI (Single Table Inheritance). We opened an issue at
Rails GitHub account: https://github.com/rails/rails/issues/28922, in
order to know how to deal with this code.

While this is fixed some other changes are being applied.

https://trello.com/c/VfNa5REK/