Refactoring on models with testing passing as expected
After reading and discussing a lot about the best way to deal with
polymorphism and single table inheritance for the case of `User`s and
its children `Producer` and `Warehouse`, and the way those managed their
places, I've managed to refactor the code using the proper migrations
(up and down cases), as well as fixing the models and the controllers to
have the expected tests passing as they were before this change.

This guarantees the application still works as outsiders expect, but
having a much cleaner code.

https://trello.com/c/VfNa5REK