Finishing authentication and login
With https://github.com/annipi have worked in fixing the authentication
system as well as defining strategies to keep a session alive (only one
month). Since this change affected how the database is represented at
the code side, we do not recommend using the `populate` script to
populate the DB.

We currently have 83 tests passing, precisely those for the endpoints
exposed in the apidoc endpoint.

https://trello.com/c/VfNa5REK/