Fixes on QR code, endpoints for Users and login auth
With https://github.com/annipi, we have worked on getting a nice qr code
for Producers or Warehouses to print in their Product's Packages. The qr
code will take the user to an address with no more than a hash. The
system looks for that Package up, and shows a nice summary of what have
happened with that specific package.

We have also modified the routes to avoid using Producers and Warehouses
ids, in order to protect their sensible information. The only way to
ensure retrieving one of their information is by using the
authentication system via the exposed token.

The token will live up for a month or until the user decides to end it
using the sign-out endpoint.